+++ src/index.tsx
@@ -2,6 +2,7 @@ import { definePlugin, routerHook } from "@decky/api";
 import { staticClasses } from "@decky/ui";
 import { FaVolumeUp } from "react-icons/fa";
 import { startAppWatcher, stopAppWatcher } from "./appWatcher";
+import { initUpdateFlow } from "./updateFlow";
 import { DiagnosticsPage } from "./components/DiagnosticsPage";
 import { QuickAccess } from "./components/QuickAccess";
 import { SetupPage } from "./components/SetupPage";
@@ -11,6 +12,7 @@ export default definePlugin(() => {
   routerHook.addRoute("/ally-dsp/setup", SetupPage, { exact: true });
   routerHook.addRoute("/ally-dsp/diagnostics", DiagnosticsPage, { exact: true });
   startAppWatcher();
+  initUpdateFlow();
   return {
     name: t.title,
     titleView: <div className={staticClasses.Title}>{t.title}</div>,